js func of 4
